Welfare's Synthetic Indicator ~~ p2distance ~~

The welfare's synthetic indicator (see references) provides an ideal tool for measuring multi-dimensional concepts such as welfare, development, living standards, etc. It enables information from the various indicators to be aggregated into a single synthetic measure. This synthetic indicator, also called \(P_{2}\) distance, overcomes several limitations of others traditional approach as Principal Components Analysis (PCA), Data Envelopment Analysis (DEA), etc. and offers some good statistical propierties and allows inter-spatial an inter-temporal comparisons.

Details

Package:	p2distance

Type:	Package

Version:	1.0.1

Date:	2012-05-02

License:	GPL

LazyLoad:	yes

~~ p2distance is the main function ~~

Author

A.J. Perez-Luque; R. Moreno; R. Perez-Perez and F.J. Bonet

References

Pena, J. B. (1977). Problemas de la medición del bienestar y conceptos afines (una aplicación al caso Español). Madrid: INE.

Pena, J. B. (2009). La medición del bienestar social: una revisión crítica. Estudios de Economía Aplicada, 27(2), 299–324.

Somarriba, N. (2008). Aproximación a la medición de la calidad de vida en la Unión Europea. Doctoral Thesis. University of Valladolid. http://www.eumed.net/tesis/2010/mnsa/index.htm

Zarzosa, P. (1992). Aproximación a la medición del bienestar social, estudio de la idoneidad del indicador sintético Distancia P2. Doctoral Thesis. University of Valladolid.

Zarzosa, P. (1996). Aproximación a la medición del Bienestar social. Valladolid: University of Valladolid. 248 pp.